securecore/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3
python-dotenv==1.0.1
cryptography==44.0.2
